Clean up GEAR radiation: dead code, unused params, magic constants
- Restore radiation_print (previously a no-op referencing removed
  mass_min/mass_max fields) with real content, and wire it into
  stellar_model_print alongside its sibling submodule print calls, matching
  the convention already used for the IMF/lifetime/supernovae submodules.
- Remove unused BB_NU_INTEGRATION_STEPS/RADIATION_N_IONIZATION_STEPS
  macros and unused phys_const/cosmo params on
  radiation_get_physical_IR_opacity/radiation_get_physical_IR_optical_depth.
- Name and centralize the ionizing-photon-rate table's 1e50 rescaling
  constant (previously a duplicated magic literal at 4 call sites, two with
  stale unresolved TODOs).
- Promote HII_region.end_time to double to match the double-typed API
  already used everywhere it's read/written (zero memory cost: the field
  sits in existing alignment padding after star_id).
- Add a doxygen map at HII_couple_ionization_rate's declaration listing
  every file/function that branches on it, since the flag's effects are
  spread across 6 files with no single place documenting them together.

+++ src/tracers/GEAR/tracers_struct.h
@@ -37,7 +37,7 @@ struct tracers_xpart_data {
     long long star_id;
 
     /*! Time when this particle is not ionized anymore */
-    float end_time;
+    double end_time;
 
     /*! Mean photon energy above the 13.6 eV HI ionization threshold for
         the tagging star, frozen at tag time (only set when
